requirements.txt
flask~=2.2.2
mysql-connector-python~=8.0.31
PyJWT~=2.6.0
requests~=2.28.1
Flask-Cors~=3.0.10
Flask-Bcrypt~=1.0.1
Flask-MySQL~=1.5.2
Flask-RESTful~=0.3.9